added some final touches – empty states, error states, etc.
