+++ solutions/domains-api/.env.example
@@ -1,4 +1,4 @@
 # Vercel API secrets
-AUTH_BEARER_TOKEN=
 VERCEL_PROJECT_ID=
 VERCEL_TEAM_ID=
+AUTH_BEARER_TOKEN=
