+++ solutions/domains-api/pages/api/remove-domain.js
@@ -1,5 +1,10 @@
-export default async function removeDomain(req, res) {
+export default async function handler(req, res) {
   const { domain } = req.query
+
+  // not required –> only for this demo to prevent removal of a few restricted domains
+  if (restrictedDomains.includes(domain)) {
+    return res.status(403).end()
+  }
 
   const response = await fetch(
     `https://api.vercel.com/v9/projects/${process.env.VERCEL_PROJECT_ID}/domains/${domain}?teamId=${process.env.VERCEL_TEAM_ID}`,
@@ -14,3 +19,5 @@ export default async function removeDomain(req, res) {
   const json = await response.json()
   res.status(200).send(json)
 }
+
+const restrictedDomains = ['portfolio.steventey.com', 'cat.vercel.pub']
