+++ solutions/static-tweets-tailwind/.env.example
@@ -1,2 +1,3 @@
-# Twitter OAuth Bear Token
-TWITTER_AUTH_TOKEN=
+VERCEL_PROJECT_ID=
+VERCEL_TEAM_ID=
+AUTH_BEARER_TOKEN=
