+++ solutions/static-tweets-tailwind/.gitignore
@@ -25,6 +25,7 @@ yarn-debug.log*
 yarn-error.log*
 
 # Local ENV files
+.env
 .env.local
 .env.development.local
 .env.test.local
